+++ CHANGELOG.md
@@ -10,7 +10,12 @@
 
 * Updated README.md
   
-## 0.0.3
+## 0.0.4
 
 * Added alchemy light account support for entrypoint v.0.6
 * Fixed sending batch transactions
+
+## 0.0.5
+
+* Added Polygon Amoy testnet network support
+* Added wallet estimateGasForSingleOperation and estimateGasForBatchedOperation functions to get transaction fee which will be paid by paymaster